refactor(emitter): split commonjs binding collectors
